wiki-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
httpx==0.26.0
pydantic==2.6.1
python-multipart==0.0.9 
numpy==1.26.0